Fix: GitHub/Discord navbar icons not displaying
GitHub and Discord icons in the navbar were not displaying correctly in the deployed version of the site.
The CSS minifier that runs during build couldn't parse the nested CSS 'img' selector inside the `.navbar__brand` and ` .navbar__logo` classes.
Flattened the CSS selectors so the minifier parses them correctly.

Tested this with a local build and verified that they're displaying properly now.
Wasn't catching this before because the preview / dev build process did doesn't minify CSS so it worked fine there.